+++ tests/Form.test.jsx
@@ -1,0 +1,100 @@
+import React from 'react';
+import { mount } from 'enzyme';
+import Schema from 'form-schema-validation';
+import {
+    Form,
+    TextField,
+    SubmitField
+} from '../src/components';
+
+
+describe('Form', () => {
+    it('should run error method from props',() => {
+        const loginSchema = new Schema({
+            login:{
+                type: String,
+                required: true
+            },
+            password: {
+                type: String,
+                required: true
+            }
+        });
+
+        const mockSubmit = jest.fn();
+        const mockError = jest.fn();
+
+        const wrapper = mount(
+            <Form
+                schema={loginSchema}
+                onSubmit={mockSubmit}
+                onError={mockError}
+            >
+                <TextField name="login" label="Login" type="text" />
+                <TextField name="password" label="Password" type="text" />
+                <SubmitField value="Login" />
+            </Form>
+        );
+        const fieldSubmit = wrapper.find(SubmitField);
+        fieldSubmit.find('button').simulate('click');
+        expect(mockSubmit.mock.calls.length).toBe(0);
+        expect(mockError.mock.calls.length).toBe(1);
+    });
+
+    it('should run submit method from props',() => {
+        const loginSchema = new Schema({
+            login:{
+                type: String,
+                required: true
+            },
+            password: {
+                type: String,
+                required: true
+            }
+        });
+
+        const mockSubmit = jest.fn();
+        const mockError = jest.fn();
+        const model = {
+            login: 'testLogin',
+            password: 'testPassword'
+        };
+
+        const wrapper = mount(
+            <Form
+                schema={loginSchema}
+                onSubmit={mockSubmit}
+                onError={mockError}
+                model={model}
+            >
+                <TextField name="login" label="Login" type="text" />
+                <TextField name="password" label="Password" type="text" />
+                <SubmitField value="Login" />
+            </Form>
+        );
+        const fieldSubmit = wrapper.find(SubmitField);
+        fieldSubmit.find('button').simulate('click');
+        expect(mockSubmit.mock.calls.length).toBe(1);
+        expect(mockError.mock.calls.length).toBe(0);
+    });
+
+    it('should update model on field value change',() => {
+        const model = {
+            title: 'testTitle'
+        };
+
+        const wrapper = mount(
+            <Form
+                onSubmit={({title}) => expect(title).toBe('test')}
+                model={model}
+            >
+                <TextField name="title" label="Title" />
+                <SubmitField value="Submit" />
+            </Form>
+        );
+        const fieldSubmit = wrapper.find(SubmitField);
+        const fieldTitle = wrapper.find(TextField);
+        fieldTitle.find('input').simulate('change', 'test');
+        fieldSubmit.find('button').simulate('click');
+    });
+});
